refresh토큰에 여러 사용자 정보를 담아서 사용자 계정을 전환할 수 있도록 구현 & 유효성 검사 추가

+++ middlewares/auth-middleware.js
@@ -2,11 +2,10 @@ const jwt = require("jsonwebtoken");
 const { Users, Tokens } = require("../models");
 
 module.exports = async (req, res, next) => {
+  const existReFreshToken = await Tokens.findOne({ order: [["createdAt", "DESC"]] });
+  const { accessToken } = req.cookies;
+  const [accessAuthType, accessAuthToken] = (accessToken ?? "").split(" ");
   try {
-    const existReFreshToken = await Tokens.findOne({});
-    const { accessToken } = req.cookies;
-    const [accessAuthType, accessAuthToken] = (accessToken ?? "").split(" ");
-
     // 1) accessToken과 refreshToken이 둘다 없을때
     if (accessAuthType !== "Bearer" && !accessAuthToken && !existReFreshToken) {
       res.status(403).json({
@@ -16,8 +15,10 @@ module.exports = async (req, res, next) => {
       return;
     }
 
-    // 2) refreshToken만 있을 때
+    // 2) refreshToken들만 있을 때
     if (existReFreshToken.tokenId.length !== 0 && !accessAuthType && !accessAuthToken) {
+      jwt.verify(existReFreshToken.tokenId, process.env.JWT_SECRET_KEY);
+
       const accessToken = jwt.sign(
         { userId: existReFreshToken.UserId },
         process.env.JWT_SECRET_KEY,
@@ -39,27 +40,66 @@ module.exports = async (req, res, next) => {
       res.locals.userNickname = user.nickname;
       next();
     } else {
-      // 4) 둘 다 있을 때
-      const { userId } = jwt.verify(accessAuthToken, process.env.JWT_SECRET_KEY);
-      const user = await Users.findOne({ where: { userId } });
+      try {
+        // 3)
+        const { userId } = jwt.verify(accessAuthToken, process.env.JWT_SECRET_KEY);
+        const user = await Users.findOne({ where: { userId } });
 
-      if (!user) {
-        res.clearCookie("accessToken");
-        res.status(403).json({ success: false, errorMessage: "토큰 사용자가 존재하지 않습니다." });
+        if (!user) {
+          res.clearCookie("accessToken");
+          res
+            .status(403)
+            .json({ success: false, errorMessage: "토큰 사용자가 존재하지 않습니다." });
+        }
+
+        res.locals.user = user;
+        res.locals.userNickname = user.nickname;
+
+        next();
+      } catch (error) {
+        // 둘 다 있는데 accessToken만 만료
+        if (error.name === "TokenExpiredError") {
+          jwt.verify(existReFreshToken.tokenId, process.env.JWT_SECRET_KEY);
+
+          const accessToken = jwt.sign(
+            { userId: existReFreshToken.UserId },
+            process.env.JWT_SECRET_KEY,
+            {
+              expiresIn: "1h",
+            }
+          );
+          res.cookie("accessToken", `Bearer ${accessToken}`);
+          const { userId } = jwt.verify(accessToken, process.env.JWT_SECRET_KEY);
+          const user = await Users.findOne({ where: { userId } });
+
+          if (!user) {
+            res.clearCookie("accessToken");
+            res
+              .status(403)
+              .json({ success: false, errorMessage: "토큰 사용자가 존재하지 않습니다." });
+          }
+
+          res.locals.user = user;
+          res.locals.userNickname = user.nickname;
+          next();
+        }
       }
-
-      res.locals.user = user;
-      res.locals.userNickname = user.nickname;
-
-      next();
     }
   } catch (error) {
-    res.clearCookie("accessToken");
-    Tokens.destroy({});
-    res.status(403).json({
-      success: false,
-      errorMessage: "전달된 쿠키에서 오류가 발생하였습니다.",
-    });
-    return;
+    if (error.name === "TokenExpiredError") {
+      res.status(400).json({
+        success: false,
+        message: "토큰이 만료된 아이디입니다. 다시 로그인 해주세요.",
+      });
+      return;
+    } else {
+      res.clearCookie("accessToken");
+      Tokens.destroy({});
+      res.status(403).json({
+        success: false,
+        errorMessage: "전달된 쿠키에서 오류가 발생하였습니다.",
+      });
+      return;
+    }
   }
 };
